Move pytorch coupling code to preconditioners/
